Relocate feather metadata loader to plugin resources
Move src/plugin/src/feather/metadata.js into src/plugin/src/resources/feather-metadata.js so the Feather JSON loader sits alongside other resource helpers. Update the transformer and tests to import the new path to keep build wiring intact.
