+++ src/plugin/src/resources/feather-metadata.js
